+++ src/entities/filters/index.ts
@@ -1,0 +1,2 @@
+export { FilterProvider } from './model/filter-provider.tsx';
+export { FilterContext, Filters, useFilter } from './model/filter-slice.ts';
