+++ src/features/new-task-field/index.ts
@@ -1,1 +1,1 @@
-export * from './ui';
+export { NewTaskField } from './ui/new-task-field.tsx';
