--- src/features/new-task-field/ui/index.ts
@@ -1,1 +1,0 @@
-export * from './new-task-field.tsx';
