--- src/features/tasks-controls/model/index.ts
@@ -1,2 +1,0 @@
-export * from './filter-provider.tsx';
-export * from './filter-slice.ts';
