--- src/features/tasks-controls/ui/index.ts
@@ -1,1 +1,0 @@
-export * from './tasks-controls.tsx';
