+++ src/features/tasks-list/index.ts
@@ -1,1 +1,1 @@
-export * from './ui';
+export { TasksList } from './ui/tasks-list.tsx';
