--- src/features/tasks-list/ui/index.ts
@@ -1,1 +1,0 @@
-export * from './tasks-list.tsx';
