--- src/shared/tests/helpers/index.ts
@@ -1,2 +1,0 @@
-export * from './render-with-context.tsx';
-export * from './setup.ts';
