+++ src/shared/tests/helpers/render-with-context.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
-import { FilterProvider } from 'src/features/tasks-controls';
-import { Task, TasksProvider } from 'src/widgets/todos';
+import { FilterProvider } from 'src/entities/filters';
+import { Task, TasksProvider } from 'src/entities/tasks';
 import React from 'react';
 
 export const renderWithContext = (component: React.ReactNode, initialTasks?: Array<Task>) => {
