+++ src/shared/tests/index.ts
@@ -1,2 +1,2 @@
-export * from './helpers';
-export * from './setupTests.ts';
+export { renderWithContext } from './helpers/render-with-context.tsx';
+export { setup } from './helpers/setup.ts';
