--- src/widgets/todos/model/index.ts
@@ -1,2 +1,0 @@
-export * from './tasks-slice.ts';
-export * from './tasks-provider.tsx';
