--- src/widgets/todos/ui/index.ts
@@ -1,1 +1,0 @@
-export * from 'src/widgets/todos/ui/todos-widget.tsx';
